+++ android/app/build.gradle
@@ -30,6 +30,11 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+
+//    buildFeatures {
+//        // 뷰 바인딩 활성화
+//        viewBinding true
+//    }
 }
 
 // Download tflite model
