+++ android/app/src/main/AndroidManifest.xml
@@ -8,9 +8,9 @@
         android:required="true" />
     <application
         android:allowBackup="true"
-        android:icon="@drawable/ic_launcher"
+        android:icon="@drawable/letter_h"
         android:label="@string/tfe_pe_app_name"
-        android:roundIcon="@drawable/ic_launcher"
+        android:roundIcon="@drawable/letter_h"
         android:supportsRtl="true"
         android:theme="@style/Theme.PoseEstimation">
         <activity android:name=".MainActivity">
@@ -19,6 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".SubActivity2"> </activity>
     </application>
 
 </manifest>
