+++ android/app/src/main/res/drawable/side_kick.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:tools="http://schemas.android.com/tools" xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:drawable="@drawable/side_kick"
+        android:width="300dp"
+        android:height="300dp"
+        tools:ignore="ResourceCycle" />
+</selector>
