+++ android/app/src/main/res/layout/activity_main.xml
@@ -1,65 +1,89 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <SurfaceView
-        android:id="@+id/surfaceView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <include layout="@layout/bottom_sheet_layout"/>
+    tools:context=".SubActivity2">
 
     <TextView
-        android:id="@+id/textView"
-        android:layout_width="398dp"
-        android:layout_height="wrap_content"
-        android:text="Exerlearning"
-        android:textAlignment="center"
-        android:textColor="@color/purple_500"
-        android:textSize="40sp"
-        android:translationX="30dp"
-        android:translationY="20dp" />
-
-    <TextView
-        android:id="@+id/countLayout"
-        android:layout_width="388dp"
-        android:layout_height="480dp"
-        android:text="Count : "
-        android:textAlignment="center"
-        android:textColor="#FFFF0000"
-        android:textScaleX="1"
-        android:textSize="50dp"
-        android:translationY="50dp"
-        app:layout_anchor="@+id/surfaceView"
-        app:layout_anchorGravity="center" />
-
-    <Button
-        android:id="@+id/button_retry"
+        android:id="@+id/textView2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Retry"
-        android:textColor="@android:color/holo_green_light"
-        android:textSize="24sp"
-        android:translationY="600dp" />
+        android:layout_marginTop="48dp"
+        android:text="@string/exerlearning"
+        android:textAppearance="@style/TextAppearance.AppCompat.Large"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="102dp"
-        android:layout_height="89dp"
-        app:srcCompat="@drawable/inu" />
 
-    <TextView
-        android:id="@+id/percent"
-        android:layout_width="180dp"
-        android:layout_height="102dp"
-        android:text="11"
-        android:textColor="#7AD80D"
-        android:textSize="60sp"
-        android:translationX="120dp"
-        android:translationY="475dp" />
+    <Button
+        android:id="@+id/btn_side_kick"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="154dp"
+        android:layout_marginTop="537dp"
+        android:layout_marginEnd="139dp"
+        android:layout_marginBottom="124dp"
+        android:text="@string/side_kick"
 
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView2" />
+
+    <ScrollView
+        android:id="@+id/ScollView"
+        android:layout_width="376dp"
+        android:layout_height="154dp"
+        android:layout_marginStart="19dp"
+        android:layout_marginTop="356dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="3dp"
+        app:layout_constraintBottom_toTopOf="@+id/btn_side_kick"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView2"
+        app:layout_constraintVertical_bias="0.0"
+        tools:ignore="SpeakableTextPresentCheck">
+
+        <LinearLayout
+            android:id="@+id/LinearLayout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <ImageButton
+                android:id="@+id/btn_side_kick_image"
+                android:layout_width="150dp"
+                android:layout_height="150dp"
+                android:background="#00FFFFFF"
+                android:scaleType="fitCenter"
+                android:contentDescription="side kick image"
+                app:srcCompat="@drawable/side_kick" />
+
+            <ImageButton
+                android:id="@+id/btn_example_1"
+                android:layout_width="150dp"
+                android:layout_height="150dp"
+                android:background="#00FFFFFF"
+                android:scaleType="fitCenter"
+                app:srcCompat="@drawable/example_image1" />
+
+            <ImageButton
+                android:id="@+id/btn_example_2"
+                android:layout_width="150dp"
+                android:layout_height="150dp"
+                android:background="#00FFFFFF"
+                android:scaleType="fitCenter"
+                app:srcCompat="@drawable/example_image2" />
+
+
+        </LinearLayout>
+    </ScrollView>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
