+++ android/local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sat Apr 30 19:24:28 KST 2022
-sdk.dir=C\:\\Users\\kmmjj\\AppData\\Local\\Android\\Sdk
+#Sat Jun 04 14:51:02 KST 2022
+sdk.dir=/Users/Soi/Library/Android/sdk
